fix(binary-formats): align OTB/OTBM struct types to uint32_t and fix buffer deallocations
- Replace 64-bit platform-dependent unsigned long with fixed-width uint32_t in itemloader.h (flags_t, VERSIONINFO) to fix corrupted items.otb binary reading.
- Fix OTBM_root_header in iomapotbm.cpp to enforce exact 16-byte struct size on 64-bit systems.
- Optimize PropStream::GET_STRING using ret.assign(p, str_len) to eliminate heap allocations and mismatching delete operator.
- Fix array deallocation delete[] for buffer resizing and cache blocks in fileloader.cpp.
- Remove invalid xmlFreeOTSERV(nodeValue) on NULL pointers in spawn.cpp.

+++ source/items.cpp
@@ -33,9 +33,9 @@ extern LuaScript g_config;
 
 extern Spells spells;
 
-long Items::dwMajorVersion = 0;
-long Items::dwMinorVersion = 0;
-long Items::dwBuildNumber = 0;
+uint32_t Items::dwMajorVersion = 0;
+uint32_t Items::dwMinorVersion = 0;
+uint32_t Items::dwBuildNumber = 0;
 
 ItemType::ItemType()
 {
@@ -240,7 +240,7 @@ int Items::loadFromOtb(std::string file)
 		//4 byte flags
 		//attributes
 		//0x01 = version data
-		unsigned long flags;
+		uint32_t flags;
 		if(!props.GET_ULONG(flags)){
 			return ERROR_INVALID_FORMAT;
 		}
